Household post and get endpoints

+++ database/migrations/2016_10_29_113441_create_household_table.php
@@ -16,8 +16,8 @@ class CreateHouseholdTable extends Migration
         Schema::create('household', function (Blueprint $table) {
             $table->increments('id');
             $table->timestamps();
-            $table->float('coordinatesX');
-            $table->float('coordinatesY');
+            $table->decimal('coordinatesX',12,10);
+            $table->decimal('coordinatesY',12,10);
             $table ->string('imagePathMain');
             $table->string('imagePathThumbnail');
         });
